Separate interface from implementation
