+++ device/api/umd/device/tt_device/hang_detection/blackhole_hang_detector.hpp
@@ -7,7 +7,7 @@
 
 #include <cstdint>
 
-#include "umd/device/tt_device/hang_detection/hang_detector.hpp"
+#include "umd/device/tt_device/hang_detection/hang_detector_implementation.hpp"
 #include "umd/device/types/xy_pair.hpp"
 
 namespace tt::umd {
@@ -18,7 +18,7 @@ enum class NocId : uint8_t;
 // Blackhole variant: reads BAR and NOC node ID from the PCIe tile.
 // NOC1 hang detection requires knowing whether NOC translation is enabled,
 // because the PCIe tile y-coordinate differs (0 vs 11).
-class BlackholeHangDetector : public HangDetector {
+class BlackholeHangDetector : public HangDetectorImplementation {
 public:
     BlackholeHangDetector(
         DeviceProtocol* protocol, architecture_implementation* arch_impl, bool noc_translation_enabled);
