+++ device/api/umd/device/tt_device/hang_detection/hang_detector.hpp
@@ -6,25 +6,12 @@
 #pragma once
 
 #include <cstdint>
-#include <functional>
 #include <optional>
 
 #include "umd/device/arch/architecture_implementation.hpp"
 #include "umd/device/types/noc_id.hpp"
-#include "umd/device/types/xy_pair.hpp"
 
 namespace tt::umd {
-
-class DeviceProtocol;
-class PcieInterface;
-enum class NocId : uint8_t;
-
-/**
- * Reads a 32-bit NOC register at (core, addr) over the given NOC and returns its value.
- * Injection point for the HangDetector NOC liveness read; defaults to a DeviceProtocol-based
- * reader and can be overridden via set_noc_reg_reader().
- */
-using NocRegReader = std::function<uint32_t(tt_xy_pair core, uint64_t addr, NocId noc)>;
 
 /**
  * HangDetector checks whether the device hardware is hung.
@@ -47,33 +34,12 @@ public:
     std::optional<bool> is_bus_hung(uint32_t data_read = HANG_READ_VALUE);
     std::optional<bool> is_noc_hung(NocId noc);
 
-    // Overrides the NOC liveness register reader (see NocRegReader). An empty function is ignored,
-    // so the default protocol-based reader stays in place.
-    void set_noc_reg_reader(NocRegReader reader);
-
 protected:
-    HangDetector(DeviceProtocol* protocol, architecture_implementation* arch_impl);
-
-    DeviceProtocol* get_protocol() const { return protocol_; }
-
-    PcieInterface* get_pcie_interface() const { return pcie_interface_; }
-
-    architecture_implementation* get_arch_impl() const { return arch_impl_; }
-
-    // Reads a 32-bit NOC register through the configured NocRegReader. Arch-specific variants compute the
-    // (core, addr) for their hang-check register and route the actual read through here.
-    uint32_t read_noc_reg(tt_xy_pair core, uint64_t addr, NocId noc);
-
     // Arch-specific implementations.
     virtual uint32_t read_hang_check_reg_via_bar() = 0;
     virtual uint32_t read_hang_check_reg_via_noc(NocId noc) = 0;
-
-private:
-    DeviceProtocol* protocol_;
-    PcieInterface* pcie_interface_;
-    bool is_mmio_protocol_;
-    architecture_implementation* arch_impl_;
-    NocRegReader noc_reg_reader_;
+    virtual bool is_bus_available() = 0;
+    virtual bool is_noc_available() = 0;
 };
 
 }  // namespace tt::umd
