+++ device/api/umd/device/tt_device/hang_detection/wormhole_hang_detector.hpp
@@ -7,7 +7,7 @@
 
 #include <cstdint>
 
-#include "umd/device/tt_device/hang_detection/hang_detector.hpp"
+#include "umd/device/tt_device/hang_detection/hang_detector_implementation.hpp"
 #include "umd/device/types/xy_pair.hpp"
 
 namespace tt::umd {
@@ -16,7 +16,7 @@ class architecture_implementation;
 enum class NocId : uint8_t;
 
 // Wormhole variant: reads BAR and NOC node ID from the ARC tile.
-class WormholeHangDetector : public HangDetector {
+class WormholeHangDetector : public HangDetectorImplementation {
 public:
     WormholeHangDetector(DeviceProtocol* protocol, architecture_implementation* arch_impl);
